extract functionality of command parsing into Parser.java class, extract functionality of UI from all other classes into SirDuke.java, add Main class
