+++ src/main/java/Main.java
@@ -1,0 +1,7 @@
+import frontend.SirDuke;
+public class Main {
+    public static void main(String[] args) {
+        SirDuke chatBot = new SirDuke();
+        chatBot.start();
+    }
+}
